+++ src/Ext/TechnoType/Body.h
@@ -312,6 +312,8 @@ public:
 	Nullable<bool> AttackMove_Aggressive;
 	Nullable<bool> AttackMove_UpdateTarget;
 
+	Nullable<bool> ApproachTarget_StopWhenInRange;
+
 	Valueable<bool> BunkerableAnyway;
 	Valueable<bool> KeepTargetOnMove;
 	Valueable<int> KeepTargetOnMove_Weapon;
@@ -706,6 +708,8 @@ public:
 
 		, AttackMove_Aggressive {}
 		, AttackMove_UpdateTarget {}
+
+		, ApproachTarget_StopWhenInRange {}
 
 		, BunkerableAnyway { false }
 		, KeepTargetOnMove { false }